Remove transformers
